Allow deletion of backups where the service is None
If a backup is created while both cinder-backup and rabbitmq
are not running, backup records are created in the database with a
service entry of None.

There is no actual backup data created on the service, since the service
isn't running. This fix allows removal of these records without an error
when a delete backup request is received.

This fix also ensures backup status is set to error in the event of
an exception during a delete.

Fixes bug #1162908

+++ cinder/backup/manager.py
@@ -238,23 +238,26 @@ class BackupManager(manager.SchedulerDependentManager):
             raise exception.InvalidBackup(reason=err)
 
         backup_service = backup['service']
-        configured_service = FLAGS.backup_service
-        if backup_service != configured_service:
-            err = _('delete_backup aborted, the backup service currently'
-                    ' configured [%(configured_service)s] is not the'
-                    ' backup service that was used to create this'
-                    ' backup [%(backup_service)s]') % locals()
-            self.db.backup_update(context, backup_id, {'status': 'available'})
-            raise exception.InvalidBackup(reason=err)
+        if backup_service is not None:
+            configured_service = FLAGS.backup_service
+            if backup_service != configured_service:
+                err = _('delete_backup aborted, the backup service currently'
+                        ' configured [%(configured_service)s] is not the'
+                        ' backup service that was used to create this'
+                        ' backup [%(backup_service)s]') % locals()
+                self.db.backup_update(context, backup_id,
+                                      {'status': 'error'})
+                raise exception.InvalidBackup(reason=err)
 
-        try:
-            backup_service = self.service.get_backup_service(context)
-            backup_service.delete(backup)
-        except Exception as err:
-            with excutils.save_and_reraise_exception():
-                self.db.backup_update(context, backup_id, {'status': 'error',
-                                                           'fail_reason':
-                                                           unicode(err)})
+            try:
+                backup_service = self.service.get_backup_service(context)
+                backup_service.delete(backup)
+            except Exception as err:
+                with excutils.save_and_reraise_exception():
+                    self.db.backup_update(context, backup_id,
+                                          {'status': 'error',
+                                           'fail_reason':
+                                           unicode(err)})
 
         context = context.elevated()
         self.db.backup_destroy(context, backup_id)
